pyproject.toml
[project]
name = "projeto-ml"
version = "0.1.0"
description = ""
authors = [
    {name = "Lucas Gabriel",email = "lucasglsilva7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib (>=3.10.1,<4.0.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "azure-ai-projects (>=1.0.0b8,<2.0.0)",
    "azure-identity (>=1.21.0,<2.0.0)",
    "openai (>=1.74.0,<2.0.0)",
    "azure-ai-inference (>=1.0.0b9,<2.0.0)",
    "azure-search-documents (>=11.5.2,<12.0.0)",
    "azure-monitor-opentelemetry (>=1.6.7,<2.0.0)",
    "transformers (>=4.51.3,<5.0.0)",
    "azure-ai-textanalytics (>=5.3.0,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "azure-ai-formrecognizer (>=3.3.3,<4.0.0)",
    "azure-core (>=1.33.0,<2.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "google-genai (>=1.12.1,<2.0.0)",
    "google-generativeai (>=0.3.2,<0.4.0)"
]

[tool.poetry]
packages = [{include = "projeto_ml", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
